requirements.txt
Flask
pytesseract
pdf2image
Pillow
dateparser
python-dateutil
spacy
